pyproject.toml
[project]
name = "fastapi-docswhatsapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "fastapi (>=0.116.2,<0.117.0)",
    "uvicorn (>=0.36.0,<0.37.0)",
    "python-multipart>=0.0.6",
    "reportlab>=4.0.4",
    "pillow>=10.0.0",
    "google-generativeai>=0.3.0",
    "supabase>=2.3.0",
    "openpyxl>=3.1.2",
    "python-dotenv>=1.0.0",
    "aiofiles>=23.2.1",
    "pydantic-settings>=2.0.0",
    "weasyprint (>=66.0,<67.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
